# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/GetTicketInfoA": {
            "post": {
                "description": "Respond with Ticket info.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get token string",
                        "name": "BearerToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "GetTicketInfo input body",
                        "name": "GetTicketInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.GetTicketInfo_input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.GetTicketInfo"
                        }
                    },
                    "400": {
                        "description": "Empty Body"
                    },
                    "401": {
                        "description": "token in not valid or role is not admin"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/GetTicketInfoU": {
            "get": {
                "description": "Respond with User ticket info.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get token string",
                        "name": "BearerToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "GetTicketInfo input body",
                        "name": "GetTicketInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.GetTicketInfo_input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.GetTicketInfo"
                        }
                    },
                    "400": {
                        "description": "Empty Body"
                    },
                    "401": {
                        "description": "Token is not valid"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/getTicketsListA": {
            "get": {
                "description": "Respond with admins tickets list.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get token string",
                        "name": "BearerToken",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.GetTicketsList_A"
                        }
                    },
                    "401": {
                        "description": "token in not valid or role is not admin"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/getTicketsListU": {
            "get": {
                "description": "Respond with users tickets list.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get token string",
                        "name": "BearerToken",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.GetTicketsList_U"
                        }
                    },
                    "401": {
                        "description": "Token is not valid"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/messageForTicketU": {
            "post": {
                "description": "Respond with Json body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get token string",
                        "name": "BearerToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "MessageForTicketing input body",
                        "name": "MessageForTicketing",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.MessageForTicketing_input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesfully message saved"
                    },
                    "400": {
                        "description": "Empty Body"
                    },
                    "401": {
                        "description": "token in not valid or role is not admin"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/ticketing": {
            "post": {
                "description": "Get user's Message.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Message was successfully received"
                    },
                    "400": {
                        "description": "Empty Body"
                    },
                    "401": {
                        "description": "Null token"
                    },
                    "500": {
                        "description": "Problem in saving message"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.GetTicketInfo": {
            "type": "object",
            "properties": {
                "textTime": {
                    "type": "string"
                },
                "textsForTicket": {
                    "type": "string"
                },
                "ticketMessage": {
                    "type": "string"
                },
                "ticketSubject": {
                    "type": "string"
                }
            }
        },
        "controllers.GetTicketInfo_input": {
            "type": "object",
            "required": [
                "ticketid"
            ],
            "properties": {
                "ticketid": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
        "controllers.GetTicketsList_A": {
            "type": "object",
            "properties": {
                "hasbeenRead": {
                    "type": "integer"
                },
                "ticketId": {
                    "type": "integer"
                },
                "ticketsMessage": {
                    "type": "string"
                },
                "ticketsSubject": {
                    "type": "string"
                }
            }
        },
        "controllers.GetTicketsList_U": {
            "type": "object",
            "properties": {
                "hasbeenRead": {
                    "type": "integer"
                },
                "ticketID": {
                    "type": "integer"
                },
                "ticketMessage": {
                    "type": "string"
                },
                "ticketSubject": {
                    "type": "string"
                }
            }
        },
        "controllers.MessageForTicketing_input": {
            "type": "object",
            "required": [
                "text",
                "ticketid"
            ],
            "properties": {
                "text": {
                    "type": "string"
                },
                "ticketid": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
...
    }
}